Updating Map Generator and Week 2-4 folders
Updated the installation instructions for the map generator to improve easy of installation. Updated config and map to account for new arena size and allow the GUI to fit on most screens. Updated EKF instructions to include functions to update the states of the robot and/or markers and prevent issues with the robot's and marker's poses not being updated correctly